new derived score instead of modification to rmse
defined in a new type now, instead of modifying the original rmse
