+++ rampwf/score_types/rmse.py
@@ -7,14 +7,9 @@ class RMSE(BaseScoreType):
     minimum = 0.0
     maximum = float('inf')
 
-    def __init__(self, name='rmse', precision=2, periodicity=-1):
+    def __init__(self, name='rmse', precision=2):
         self.name = name
         self.precision = precision
-        self.periodicity = -1
 
     def __call__(self, y_true, y_pred):
-        d = np.abs(y_true - y_pred)
-        if(self.periodicity > 0):
-            d = min(np.mod(d, self.periodicity),
-                    np.mod(-d, self.periodicity))
-        return np.sqrt(np.mean(np.square(d)))
+        return np.sqrt(np.mean(np.square(y_true - y_pred)))
